fix: Moved dev dependencies to `dev_deps.ts`
