+++ .github/workflows/check.yml
@@ -24,7 +24,7 @@ jobs:
           path: |
             ~/.deno
             ~/.cache/deno
-          key: ${{ runner.os }}-deno-${{ hashFiles('**/deps.ts') }}
+          key: ${{ runner.os }}-deno-${{ hashFiles('**/*deps.ts') }}
       - name: Run tests
         run: deno task test --doc
   
@@ -51,7 +51,7 @@ jobs:
           path: |
             ~/.deno
             ~/.cache/deno
-          key: ${{ runner.os }}-deno-${{ hashFiles('**/deps.ts') }}
+          key: ${{ runner.os }}-deno-${{ hashFiles('**/*deps.ts') }}
       - name: Run tests
         run: deno task coverage
       - name: Upload lcov file
