+++ deps.ts
@@ -1,5 +1,3 @@
-// Standard library
-export * as asserts from "https://deno.land/std@0.152.0/testing/asserts.ts#^";
 export * as yaml from "https://deno.land/std@0.152.0/encoding/yaml.ts#^";
 export * as path from "https://deno.land/std@0.152.0/path/mod.ts#^";
 export * as flags from "https://deno.land/std@0.152.0/flags/mod.ts#^";
