+++ dev_deps.ts
@@ -1,0 +1,1 @@
+export * as asserts from "https://deno.land/std@0.151.0/testing/asserts.ts#^";
