+++ test/extractor/env.test.ts
@@ -1,4 +1,4 @@
-import { asserts } from "../../deps.ts";
+import { asserts } from "../../dev_deps.ts";
 import { EnvExtractor } from "../../mod.ts";
 
 Deno.test(
